Bug Fix 11282: config.getini returns an empty list for an option of type string absent in INI file
